+++ src/realm/table_view.hpp
@@ -785,7 +785,7 @@ inline bool TableViewBase::is_attached() const noexcept
 
 inline bool TableViewBase::is_row_attached(size_t row_ndx) const noexcept
 {
-    return size_t(m_row_indexes.get(row_ndx)) != detached_ref;
+    return m_row_indexes.get(row_ndx) != detached_ref;
 }
 
 inline size_t TableViewBase::size() const noexcept
@@ -1030,7 +1030,7 @@ inline int64_t TableViewBase::get_int(size_t column_ndx, size_t row_ndx) const n
 {
     REALM_ASSERT_INDEX(column_ndx, row_ndx);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_int(column_ndx, real_ndx);
 }
@@ -1039,7 +1039,7 @@ inline bool TableViewBase::get_bool(size_t column_ndx, size_t row_ndx) const noe
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Bool);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_bool(column_ndx, real_ndx);
 }
@@ -1048,7 +1048,7 @@ inline OldDateTime TableViewBase::get_olddatetime(size_t column_ndx, size_t row_
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_OldDateTime);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_olddatetime(column_ndx, real_ndx);
 }
@@ -1057,7 +1057,7 @@ inline Timestamp TableViewBase::get_timestamp(size_t column_ndx, size_t row_ndx)
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Timestamp);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_timestamp(column_ndx, real_ndx);
 }
@@ -1066,7 +1066,7 @@ inline float TableViewBase::get_float(size_t column_ndx, size_t row_ndx) const n
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Float);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_float(column_ndx, real_ndx);
 }
@@ -1075,7 +1075,7 @@ inline double TableViewBase::get_double(size_t column_ndx, size_t row_ndx) const
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Double);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_double(column_ndx, real_ndx);
 }
@@ -1084,7 +1084,7 @@ inline StringData TableViewBase::get_string(size_t column_ndx, size_t row_ndx) c
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_String);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_string(column_ndx, real_ndx);
 }
@@ -1093,7 +1093,7 @@ inline BinaryData TableViewBase::get_binary(size_t column_ndx, size_t row_ndx) c
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Binary);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_binary(column_ndx, real_ndx);
 }
@@ -1102,7 +1102,7 @@ inline Mixed TableViewBase::get_mixed(size_t column_ndx, size_t row_ndx) const n
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Mixed);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_mixed(column_ndx, real_ndx);
 }
@@ -1111,7 +1111,7 @@ inline DataType TableViewBase::get_mixed_type(size_t column_ndx, size_t row_ndx)
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Mixed);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_mixed_type(column_ndx, real_ndx);
 }
@@ -1120,7 +1120,7 @@ inline size_t TableViewBase::get_subtable_size(size_t column_ndx, size_t row_ndx
 {
     REALM_ASSERT_INDEX_AND_TYPE_TABLE_OR_MIXED(column_ndx, row_ndx);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_subtable_size(column_ndx, real_ndx);
 }
@@ -1129,7 +1129,7 @@ inline size_t TableViewBase::get_link(size_t column_ndx, size_t row_ndx) const n
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Link);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_link(column_ndx, real_ndx);
 }
@@ -1153,7 +1153,7 @@ inline bool TableViewBase::is_null_link(size_t column_ndx, size_t row_ndx) const
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Link);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->is_null_link(column_ndx, real_ndx);
 }
@@ -1422,7 +1422,7 @@ inline ConstTableView ConstTableView::find_all_olddatetime(size_t column_ndx, Ol
 inline TableView::RowExpr TableView::get(size_t row_ndx) noexcept
 {
     REALM_ASSERT_ROW(row_ndx);
-    size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get(real_ndx);
 }
@@ -1430,7 +1430,7 @@ inline TableView::RowExpr TableView::get(size_t row_ndx) noexcept
 inline TableView::ConstRowExpr TableView::get(size_t row_ndx) const noexcept
 {
     REALM_ASSERT_ROW(row_ndx);
-    size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get(real_ndx);
 }
@@ -1438,7 +1438,7 @@ inline TableView::ConstRowExpr TableView::get(size_t row_ndx) const noexcept
 inline ConstTableView::ConstRowExpr ConstTableView::get(size_t row_ndx) const noexcept
 {
     REALM_ASSERT_ROW(row_ndx);
-    size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get(real_ndx);
 }
@@ -1500,7 +1500,7 @@ inline TableRef TableView::get_subtable(size_t column_ndx, size_t row_ndx)
 {
     REALM_ASSERT_INDEX_AND_TYPE_TABLE_OR_MIXED(column_ndx, row_ndx);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_subtable(column_ndx, real_ndx);
 }
@@ -1509,7 +1509,7 @@ inline ConstTableRef TableView::get_subtable(size_t column_ndx, size_t row_ndx) 
 {
     REALM_ASSERT_INDEX_AND_TYPE_TABLE_OR_MIXED(column_ndx, row_ndx);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_subtable(column_ndx, real_ndx);
 }
@@ -1518,7 +1518,7 @@ inline ConstTableRef ConstTableView::get_subtable(size_t column_ndx, size_t row_
 {
     REALM_ASSERT_INDEX_AND_TYPE_TABLE_OR_MIXED(column_ndx, row_ndx);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->get_subtable(column_ndx, real_ndx);
 }
@@ -1527,7 +1527,7 @@ inline void TableView::clear_subtable(size_t column_ndx, size_t row_ndx)
 {
     REALM_ASSERT_INDEX_AND_TYPE_TABLE_OR_MIXED(column_ndx, row_ndx);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     return m_table->clear_subtable(column_ndx, real_ndx);
 }
@@ -1540,7 +1540,7 @@ inline void TableView::set_int(size_t column_ndx, size_t row_ndx, int64_t value)
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Int);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_int(column_ndx, real_ndx, value);
 }
@@ -1549,7 +1549,7 @@ inline void TableView::set_bool(size_t column_ndx, size_t row_ndx, bool value)
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Bool);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_bool(column_ndx, real_ndx, value);
 }
@@ -1558,7 +1558,7 @@ inline void TableView::set_olddatetime(size_t column_ndx, size_t row_ndx, OldDat
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_OldDateTime);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_olddatetime(column_ndx, real_ndx, value);
 }
@@ -1567,7 +1567,7 @@ inline void TableView::set_timestamp(size_t column_ndx, size_t row_ndx, Timestam
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Timestamp);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_timestamp(column_ndx, real_ndx, value);
 }
@@ -1576,7 +1576,7 @@ inline void TableView::set_float(size_t column_ndx, size_t row_ndx, float value)
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Float);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_float(column_ndx, real_ndx, value);
 }
@@ -1585,7 +1585,7 @@ inline void TableView::set_double(size_t column_ndx, size_t row_ndx, double valu
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Double);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_double(column_ndx, real_ndx, value);
 }
@@ -1593,7 +1593,7 @@ inline void TableView::set_double(size_t column_ndx, size_t row_ndx, double valu
 template<class E>
 inline void TableView::set_enum(size_t column_ndx, size_t row_ndx, E value)
 {
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_int(column_ndx, real_ndx, value);
 }
@@ -1602,7 +1602,7 @@ inline void TableView::set_string(size_t column_ndx, size_t row_ndx, StringData 
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_String);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_string(column_ndx, real_ndx, value);
 }
@@ -1611,7 +1611,7 @@ inline void TableView::set_binary(size_t column_ndx, size_t row_ndx, BinaryData 
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Binary);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_binary(column_ndx, real_ndx, value);
 }
@@ -1620,7 +1620,7 @@ inline void TableView::set_mixed(size_t column_ndx, size_t row_ndx, Mixed value)
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Mixed);
 
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_mixed(column_ndx, real_ndx, value);
 }
@@ -1628,7 +1628,7 @@ inline void TableView::set_mixed(size_t column_ndx, size_t row_ndx, Mixed value)
 inline void TableView::set_subtable(size_t column_ndx, size_t row_ndx, const Table* value)
 {
     REALM_ASSERT_INDEX_AND_TYPE_TABLE_OR_MIXED(column_ndx, row_ndx);
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_subtable(column_ndx, real_ndx, value);
 }
@@ -1636,7 +1636,7 @@ inline void TableView::set_subtable(size_t column_ndx, size_t row_ndx, const Tab
 inline void TableView::set_link(size_t column_ndx, size_t row_ndx, size_t target_row_ndx)
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Link);
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->set_link(column_ndx, real_ndx, target_row_ndx);
 }
@@ -1644,7 +1644,7 @@ inline void TableView::set_link(size_t column_ndx, size_t row_ndx, size_t target
 inline void TableView::nullify_link(size_t column_ndx, size_t row_ndx)
 {
     REALM_ASSERT_INDEX_AND_TYPE(column_ndx, row_ndx, type_Link);
-    const size_t real_ndx = size_t(m_row_indexes.get(row_ndx));
+    const int64_t real_ndx = m_row_indexes.get(row_ndx);
     REALM_ASSERT(real_ndx != detached_ref);
     m_table->nullify_link(column_ndx, real_ndx);
 }
